+++ src/components/common/Alert/index.tsx
@@ -1,6 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { useRecoilState } from 'recoil';
-import { AlertInnerWrapper, AlertText, AlertWrapper } from './styled';
+import {
+  AlertInner,
+  AlertInnerWrapper,
+  AlertText,
+  AlertWrapper,
+} from './styled';
 import { alertState } from '../../../store/alert';
 import './Alert.css';
 import { AnimatePresence } from 'framer-motion';
@@ -8,38 +13,40 @@ import { AnimatePresence } from 'framer-motion';
 const variants = {
   active: {
     opacity: 1,
+    scale: 1,
     y: 0,
   },
   unActive: {
     opacity: 0,
+    scale: 0,
   },
 };
 const Alert = () => {
   const [alert, setAlert] = useRecoilState(alertState);
-  const [toggle, setToggle] = useState<boolean>(true);
 
   useEffect(() => {
-    setToggle(alert.alertHandle);
     setTimeout(() => {
-      setToggle(false);
+      setAlert({
+        ...alert,
+        alertHandle: false,
+      });
     }, 4000);
   }, []);
+  console.log(alert.alertHandle);
   return (
-    <AnimatePresence>
-      {toggle && (
-        <AlertWrapper
-          variants={variants}
-          exit={'unActive'}
-          animate={'active'}
-          initial={{ opacity: 0 }}
-          transition={{ duration: 0.5 }}
-        >
-          <AlertInnerWrapper className={alert.alertStatus}>
-            <AlertText>{alert.alertMessage}</AlertText>
-          </AlertInnerWrapper>
-        </AlertWrapper>
-      )}
-    </AnimatePresence>
+    <AlertWrapper>
+      <AlertInner
+        variants={variants}
+        exit={'unActive'}
+        animate={'active'}
+        initial={{ opacity: 0, scale: 0 }}
+        transition={{ duration: 0.5 }}
+      >
+        <AlertInnerWrapper className={alert.alertStatus}>
+          <AlertText>{alert.alertMessage}</AlertText>
+        </AlertInnerWrapper>
+      </AlertInner>
+    </AlertWrapper>
   );
 };
