+++ src/components/common/input/TextInput/index.tsx
@@ -10,14 +10,16 @@ import {
 export interface Iprops {
   name?: string;
   error?: string;
+  setError?: (error: boolean) => void;
   placeholder?: string;
   image?: string;
   file?: boolean;
   onChange: (e: any) => void;
+  type?: string;
 }
 
 const TextInput = (props: Iprops) => {
-  const { name, error, placeholder, image, file, onChange } = props;
+  const { name, error, placeholder, image, file, onChange, type } = props;
 
   return (
     <>
@@ -25,6 +27,7 @@ const TextInput = (props: Iprops) => {
         <StyledInput
           className={'formInput'}
           name={name}
+          type={type}
           onChange={(e: any) => onChange(e.target.value)}
           placeholder={placeholder}
         />
