+++ src/pages/Recruit/RecruitForm/FormFunctions.ts
@@ -39,7 +39,7 @@ const checkValue = ({ type, data }: { type: string; data: string }) => {
   let validate;
   switch (type) {
     case 'phoneNumber':
-      return (validate = /^01([0|1|6|7|8|9])-?([0-9]{3,4})-?([0-9]{4})$/.test(
+      return (validate = /^01([0|1|6|7|8|9])-?([0-9]{4})-?([0-9]{4})$/.test(
         data,
       ));
     case 'email':
