+++ src/pages/Recruit/Recruitment/index.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useParams } from 'react-router-dom';
 import {
   BannerWrapper,
   ContainerInner,
@@ -7,105 +6,86 @@ import {
   NavigationBlock,
   TopMargin,
 } from '../../../styles/layouts';
-import {
-  MainText,
-  SubCategory,
-  SubTitle,
-  Title,
-} from '../../../components/common/Title/title';
-import {
-  IntroduceWrapper,
-  JoinInner,
-  JoinWrapper,
-  MoblieBottomElementWrapper,
-  MoblieTopElementWrapper,
-  SectionWrapper,
-} from './styled';
-import { recruitDefaultText } from '../../../api/pageData/recruitDefaultText';
-import { RecruitDetails } from '../../../api/pageData/recruitDetails';
-import { Banner } from '../../../img/Banner/Banner';
-import BlueBanner from '../../../img/Banner/BlueBanner.png';
-import BulletList from '../../../components/common/BulletList';
-import InformationBar from './InformationBar';
-import RecruitFaqButton from './RecruitFaqButton';
+import { MainText, Title } from '../../../components/common/Title/title';
 
-const Recruitment: React.FC = () => {
-  const { id } = useParams();
-  const aboutTeam = RecruitDetails.find((aboutTeam) => aboutTeam.id === id);
+import {
+  Category,
+  CategoryInner,
+  CategoryWrapper,
+  MobileCategoryInner,
+  MobileCategoryWrapper,
+  SubCategory,
+  TeamCategory,
+  TeamCategoryWrapper,
+  TeamSubCategory,
+} from './styled';
+import { RecruitTeam } from '../../../api/pageData/recruitTeam';
+import { useNavigate } from 'react-router';
+import { Banner } from '../../../img/Banner/Banner';
+
+import BlueBanner from '../../../img/Banner/BlueBanner.png';
+import { categoryAnimate } from '../../../components/common/Variants/Variants';
+
+const Recruitment = () => {
+  const navigate = useNavigate();
+
   return (
     <>
       <NavigationBlock />
       <BannerWrapper>
         <Banner src={BlueBanner} />
       </BannerWrapper>
-      {aboutTeam && (
-        <LayoutContainer>
-          <ContainerInner>
-            <TopMargin />
-            <Title>{aboutTeam?.name}</Title>
-            <TopMargin />
-            <MoblieTopElementWrapper>
-              <JoinInner>
-                <InformationBar name={aboutTeam.name} />
-                <TopMargin />
-              </JoinInner>
-            </MoblieTopElementWrapper>
-            <MainText>열정적인 동료를 얻기 위해 이 자리에 모였습니다.</MainText>
-            <MainText>
-              우리는 함께 고민을 나누고 도전하며 목표를 향해 달리고 있습니다.
-            </MainText>
-            <TopMargin />
-            <SectionWrapper>
-              {aboutTeam && (
-                <IntroduceWrapper>
-                  {aboutTeam.activity ? (
-                    <>
-                      <SubCategory>합류하시면 함께 할 활동입니다.</SubCategory>
-                      <BulletList text={aboutTeam.activity} />
-                      <TopMargin />
-                    </>
-                  ) : null}
-                  <SubCategory>이런 분을 찾습니다</SubCategory>
-                  <MainText>
-                    <BulletList text={recruitDefaultText.findMember} />
-                    {aboutTeam.people && <BulletList text={aboutTeam.people} />}
-                  </MainText>
-                  <TopMargin />
-                  <SubCategory>이런 경험이 있다면 더 좋습니다</SubCategory>
-                  <MainText>
-                    <BulletList text={recruitDefaultText.goodMember} />
-                    {aboutTeam.preferential && (
-                      <BulletList text={aboutTeam.preferential} />
-                    )}
-                  </MainText>
-                  <TopMargin />
-                  <SubCategory>GDSC의 혜택</SubCategory>
-                  <MainText>
-                    <BulletList text={recruitDefaultText.benefits} />
-                  </MainText>
-                  <TopMargin />
-                  <SubCategory>GDSC DJU로의 합류과정</SubCategory>
-                  <MainText>
-                    <BulletList text={recruitDefaultText.process} />
-                  </MainText>
-                </IntroduceWrapper>
-              )}
-              <JoinWrapper>
-                <JoinInner>
-                  <InformationBar name={aboutTeam.name} />
-                  <TopMargin />
-                  <RecruitFaqButton name={id as string} />
-                  <TopMargin />
-                </JoinInner>
-              </JoinWrapper>
-            </SectionWrapper>
-            <MoblieBottomElementWrapper>
-              <RecruitFaqButton name={id as string} />
-            </MoblieBottomElementWrapper>
-          </ContainerInner>
-        </LayoutContainer>
-      )}
-      <TopMargin />
+      <LayoutContainer>
+        <ContainerInner>
+          <TopMargin />
+          <Title>함께 성장할 동료를 모집합니다</Title>
+          <TopMargin />
+          <MainText>
+            열정적인 동료를 얻기 위해 이 자리에 모였습니다.
+            <br />
+            우리는 함께 고민을 나누고 도전하며 목표를 향해 달리고 있습니다.
+          </MainText>
+          <TopMargin />
+          <TeamCategoryWrapper>
+            <TeamCategory>Team</TeamCategory>
+            <TeamSubCategory>Skill</TeamSubCategory>
+            <TeamSubCategory>Stack</TeamSubCategory>
+          </TeamCategoryWrapper>
+          {/*Desk 팀 리스트*/}
+          {RecruitTeam.map((data, key) => (
+            <CategoryWrapper
+              key={key}
+              initial={'unHover'}
+              whileHover={'hovered'}
+              variants={categoryAnimate}
+              onClick={() => {
+                navigate('/recruit/detail/' + data.id);
+              }}
+            >
+              <CategoryInner>
+                <Category>{data.name}</Category>
+                <SubCategory>{data.skill}</SubCategory>
+                <SubCategory>{data.stack}</SubCategory>
+              </CategoryInner>
+            </CategoryWrapper>
+          ))}
+          {/*모바일 환경 팀 리스트*/}
+          {RecruitTeam.map((data, key) => (
+            <MobileCategoryWrapper
+              key={key}
+              onClick={() => {
+                navigate('/recruit/detail/' + data.id);
+              }}
+            >
+              <MobileCategoryInner>
+                <Category>{data.name}</Category>
+                <SubCategory>{data.skill}</SubCategory>
+              </MobileCategoryInner>
+            </MobileCategoryWrapper>
+          ))}
+          <TopMargin />
+        </ContainerInner>
+      </LayoutContainer>
     </>
   );
 };
