+++ src/pages/Recruit/RecruitmentDetail/InformationBar/index.tsx
