+++ src/pages/Recruit/RecruitmentDetail/InformationBar/styled.ts
