+++ src/pages/Recruit/index.tsx
@@ -1,14 +1,14 @@
 import React, { lazy } from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-const JoinUs = lazy(() => import('./JoinUs'));
 const Recruitment = lazy(() => import('./Recruitment'));
+const RecruitmentDetail = lazy(() => import('./RecruitmentDetail'));
 const RecruitForm = lazy(() => import('./RecruitForm'));
 const Recruit = () => {
   return (
     <Routes>
-      <Route path={'/*'} element={<JoinUs />} />
-      <Route path={'/detail/:id'} element={<Recruitment />} />
+      <Route path={'/*'} element={<Recruitment />} />
+      <Route path={'/detail/:id'} element={<RecruitmentDetail />} />
       <Route path={'/form/:id'} element={<RecruitForm />} />
     </Routes>
   );
