+++ src/pages/index.tsx
@@ -19,7 +19,7 @@ const Faq = lazy(() => import('./Faq'));
 const FaqDetail = lazy(() => import('./Faq/FaqDetail'));
 
 const Pages = () => {
-  const [loading, setLoading] = useRecoilState(loaderState);
+  const [loading] = useRecoilState(loaderState);
   return (
     <>
       {loading.load && <GoogleSpinner background={true} />}
